Added experimental support for clients using older protocol versions
This is behind the `enable_old_clients` config setting. This is enough to get
clients as old as 0.0.15a/mptest1 in-game, but it's possible to crash them e.g.
by disconecting another client, or placing an unsupported block. Newer blocks
are filtered out when sending the map, but this needs to also be done when
informing a client when a block is set during the game.