Set client socket fd to INVALID_SOCKET after closing it, and check it before closing it
Previously it was mistakenly set to 0, which is the FD for stdin. When a client
disconnected, client_disconnect set it to 0, then client_destroy would try to
close it again, inadvertently closing stdin. When a client connects after that,
FD 0 might be reused for their socket, so command_tick_readline would poll on
it thinking it's stdin, and rl_callback_read_char would swallow a byte from it,
causing the new connection to fail due to malformed data.